+++ includes/classes/ServerConfig.hpp
@@ -5,18 +5,16 @@
 #include "ServerLocation.hpp"
 #include "../webserv.hpp"
 
-//	USE SOMETHING LIFO
-typedef std::vector<ServerLocation *>	LocationBlocks;
-
-// ADD ENUM FOT LOCATION TYPES
+class ServerLocation;
+typedef std::vector<ServerLocation *>			LocationBlocks;
 
 class	ServerConfig
 {
 	public:
-		ServerConfig(void);			//	DONT FORGET TO SET DEFAULTS
-		ServerConfig(std::vector<std::string> strServerBlock);
+		ServerConfig(void);
+//		ServerConfig(std::vector<std::string> strServerBlock);
 		ServerConfig(const ServerConfig &serverConfig);
-		~ServerConfig(void);	//	DESTRUCTOR MUST DELETE POINTERS IN locationBlocks
+		~ServerConfig(void);
 
 		ServerConfig &operator = (const ServerConfig &serverConfig);
 
@@ -25,13 +23,13 @@ class	ServerConfig
 
 		//			GETTERS
 
-		uint32_t			getHost(void);
-		uint16_t			getPort(void);
-		std::string			getServerName(void);
-		size_t				getLocationBlocksSize(void);
-		ServerLocation		*getLocationFromIndex(size_t i);
-		ServerLocation		*getLocationFromPath(std::string path);
-		LOCATION_BLOCK_TYPE getLocationType(ServerLocation *location);	//	MAYBE PUT THIS IN UTILS
+		uint32_t			getHost(void) const;
+		uint16_t			getPort(void) const;
+		std::string			getServerName(void) const;
+		size_t				getLocationBlocksSize(void) const;
+		ServerLocation		*getLocationFromIndex(size_t i) const;
+		ServerLocation		*getLocationFromPath(std::string path) const;
+		LOCATION_BLOCK_TYPE getLocationType(ServerLocation *location) const;	//	MAYBE PUT THIS IN UTILS
 		//	=============================
 
 	private:
@@ -43,11 +41,3 @@ class	ServerConfig
 };
 
 std::ostream & operator << (std::ostream &out, const ServerConfig &serverConfig);
-
-typedef enum	e_location_block_type
-{
-	L_UNHANDLED,
-	L_STATIC,
-	L_REV_PROXY,
-	L_CGI
-}	LOCATION_BLOCK_TYPE;
