+++ includes/webserv.hpp
@@ -44,12 +44,7 @@ typedef std::vector<unsigned char>				ByteArr;
 
 #define CLIENT_CHUNK_SIZE						4096
 
-//TODO: REMOVE PLACEHOLDER WHEN CLASS GETS IMPLEMENTED
-struct ServerConfig
-{
-	uint32_t	ipv4;
-	uint16_t	port;
-};
+
 /******************************  CLI_COLORS  ******************************/
 #define CLI_COLORS_MAGENTA	"\033[35m"
 #define CLI_COLORS_YELLOW	"\033[33m"
