+++ sources/classes/ServerConfig.cpp
@@ -1,14 +1,12 @@
 #include "../../includes/classes/ServerConfig.hpp"
 
 	/*	CONSTRUCTORS	*/
-//	MAYBE DEFAULT VALUES COULD BE DEFINED AS MACROS?
+
 ServerConfig::ServerConfig(void)
 {
 	this->_host = Network::sToIPV4Packed("127.0.0.1");
 	this->_port = 80;
-	this->_serverName = "";
-
-	this->_rootDir = "/var/www/html";	//	THIS GOES TO LOCATIONBLOCKS
+	this->_serverName = "webserv.ft";
 
 	this->_locationBlocks.push_back(new LocationStatic);
 }
@@ -23,7 +21,7 @@ ServerConfig::ServerConfig(std::vector<std::string> strServerBlock)
 ServerConfig::ServerConfig(const ServerConfig &serverConfig)
 {
 	if (this != &serverConfig)
-		this = serverConfig;
+		*this = serverConfig;
 }
 
 ServerConfig::~ServerConfig(void)
@@ -32,78 +30,94 @@ ServerConfig::~ServerConfig(void)
 
 	nLocations = this->_locationBlocks.size();
 	for (size_t i = 0; i < nLocations; i++)
-		delete this->_locationBlocks.at(i);	//	HUH... WILL THIS WORK?
+		delete this->_locationBlocks.at(i);
 }
 
-ServerConfig ServerConfig::&operator=(const ServerConfig &serverConfig)
+ServerConfig &ServerConfig::operator=(const ServerConfig &serverConfig)
 {
 	size_t	lbSize;
 
 	this->_host = serverConfig.getHost();
 	this->_port = serverConfig.getPort();
 	this->_serverName = serverConfig.getServerName();
-	this->_maxBodySize = serverConfig.getMaxBodySize();
 
 	lbSize = serverConfig.getLocationBlocksSize();
+
+	ServerLocation	*location;
 	for (size_t i = 0; i < lbSize; i++)
-		this->_locationBlocks.push_back(new ServerLocation(serverConfig.getLocationFromIndex(i) ) );	//	DOES INHERITANCE WORK WITH THIS?
+	{
+		location = serverConfig.getLocationFromIndex(i);
+
+		switch (this->getLocationType(location))
+		{
+			case (L_STATIC):
+				this->_locationBlocks.push_back(new LocationStatic(*(dynamic_cast<LocationStatic*>(location) ) ) );
+				break ;
+			case (L_REV_PROXY):
+				break ;
+			case (L_CGI):
+			//	this->_locationBlocks.push_back(new LocationCGI(location));
+				break ;
+			case (L_UNHANDLED):
+				//	throw exception
+				break ;
+//			default:
+				//	throw exception?
+
+		}
+	}
 
 	return (*this);
 }
 
 	/*	GETTERS	*/
 
-uint32_t	ServerConfig::getHost(void)
+uint32_t	ServerConfig::getHost(void) const
 {
 	return (this->_host);
 }
 
-uint16_t	ServerConfig::getPort(void)
+uint16_t	ServerConfig::getPort(void) const
 {
 	return (this->_port);
 }
 
-std::string	ServerConfig::getServerName(void)
+std::string	ServerConfig::getServerName(void) const
 {
 	return (this->_serverName);
 }
 
-unsigned int	ServerConfig::getMaxBodySize(void)
-{
-	return (this->_maxBodySize);
-}
-
-size_t	ServerConfig::getLocationBlocksSize(void)
+size_t	ServerConfig::getLocationBlocksSize(void) const
 {
 	return (this->_locationBlocks.size());
 }
 
-ServerLocation	ServerConfig::*getLocationFromIndex(size_t i)
+ServerLocation	*ServerConfig::getLocationFromIndex(size_t i) const
 {
 	return (this->_locationBlocks.at(i));
 }
 
-ServerLocation	ServerConfig::*getLocationFromPath(std::string path)
+ServerLocation	*ServerConfig::getLocationFromPath(std::string path) const
 {
 	size_t	nLocations;
 
 	nLocations = this->_locationBlocks.size();
 	for (size_t i = 0; i < nLocations; i++)
 	{
-		if (this->_locationBlocks.at(i)->location == path)
+		if (this->_locationBlocks.at(i)->getLocation() == path)
 			return (this->_locationBlocks.at(i));
 	}
 	return (NULL);
 }
 
-LOCATION_BLOCK_TYPE	ServerConfig::getLocationType(ServerLocation *location)
+LOCATION_BLOCK_TYPE	ServerConfig::getLocationType(ServerLocation *location) const
 {
 	if (dynamic_cast<LocationStatic *>(location) != NULL)
 		return (L_STATIC);
-	else if (dynamic_cast<LocationRevProxy *>(location) != NULL)
-		return (L_REV_PROXY);
-	else if (dynamic_cast<LocationCGI *>(location) != NULL)
-		return (L_CGI);
+//	else if (dynamic_cast<LocationRevProxy *>(location) != NULL)
+//		return (L_REV_PROXY);
+//	else if (dynamic_cast<LocationCGI *>(location) != NULL)
+//		return (L_CGI);
 	else
 		return (L_UNHANDLED);
 }
@@ -114,15 +128,30 @@ std::ostream	&operator<<(std::ostream &out, const ServerConfig &serverConfig)
 
 	out << "Host: " << serverConfig.getHost() << std::endl;
 	out << "Port: " << serverConfig.getPort() << std::endl;
-	out << "Server name: " << serverConfig.getserverName() << std::endl;
-
-	out << "Max body size: " serverConfig.getmaxBodySize();	//	MOVE THIS TO LOCATIONBLOCK?
+	out << "Server name: " << serverConfig.getServerName() << std::endl;
 	
 	lbSize = serverConfig.getLocationBlocksSize();
+
+	ServerLocation	*location;
 	for (size_t i = 0; i < lbSize; i++)
 	{
-		out << "LocationBlock nr. " << i << ":" << std::endl
-		out << *(serverConfig.getLocationFromIndex(i)) << std::endl;
+		location = serverConfig.getLocationFromIndex(i);
+		out << "LocationBlock nr. " << i << ":" << std::endl;
+		//	MUST USE FUCKING DYNAMIC CAST
+		switch (serverConfig.getLocationType(location))
+		{
+			case (L_STATIC):
+				out << *(dynamic_cast<LocationStatic *>(location) ) << std::endl;
+				break ;
+			case (L_REV_PROXY):
+				break ;
+			case (L_CGI):
+				// out << *(dynamic_cast<LocationCGI *>(location) ) << std::endl;
+				break ;
+			case (L_UNHANDLED):
+				//	THROW EXCEPTION
+				break ;
+		}		
 		out << std::endl;
 	}
 	return (out);
