+++ sources/classes/ServerLocation.cpp
@@ -5,7 +5,7 @@
 ServerLocation::ServerLocation(void)
 {
 	this->_location = "/";
-	this->_rootDir = "./test_files/error_pages/"
+	this->_rootDir = "./test_files";
 	this->_indexFile = "index.html";
 	this->_maxBodySize = 1000000;
 	this->_errorPages[404] = "./test_files/error_pages/404.html";
@@ -29,17 +29,22 @@ ServerLocation::ServerLocation(const ServerLocation &serverLocation)
 		*this = serverLocation;
 }
 
-ServerLocation	ServerLocation::&operator=(const ServerLocation &serverLocation)
+ServerLocation::~ServerLocation(void)
+{
+	return ;
+}
+
+ServerLocation	&ServerLocation::operator=(const ServerLocation &serverLocation)
 {
 	this->_location = serverLocation.getLocation();
 	this->_rootDir = serverLocation.getRootDir();
 	this->_indexFile = serverLocation.getIndexFilename();
 	this->_maxBodySize = serverLocation.getMaxBodySize();
 	
-	for (IntStrMap::iterator itt = serverLocation.getErrPageIttBegin(); itt != serverLocation.getErrPageIttEnd(); itt++)
+	for (IntStrMap::const_iterator itt = serverLocation.getErrPageIttBegin(); itt != serverLocation.getErrPageIttEnd(); itt++)
 		this->_errorPages[itt->first] = itt->second;
 
-	for (IntStrMap::iterator itt = serverLocation.getRedirectionIttBegin(); itt != serverLocation.getRedirectionIttEnd(); itt++)
+	for (IntStrMap::const_iterator itt = serverLocation.getRedirectionIttBegin(); itt != serverLocation.getRedirectionIttEnd(); itt++)
 		this->_redirections[itt->first] = itt->second;
 
 	for (size_t i = 0; i < serverLocation.getMethodsAllowedSize(); i++)
@@ -48,69 +53,69 @@ ServerLocation	ServerLocation::&operator=(const ServerLocation &serverLocation)
 	return (*this);
 }
 
-std::string	ServerLocation::getLocation(void)
+std::string	ServerLocation::getLocation(void) const
 {
 	return (this->_location);
 }
 
-std::string	ServerLocation::getRootDir(void)
+std::string	ServerLocation::getRootDir(void) const
 {
 	return (this->_rootDir);
 }
 
-std::string	ServerLocation::getIndexFilename(void)
+std::string	ServerLocation::getIndexFilename(void) const
 {
 	return (this->_indexFile);
 }
 
-uint32_t	ServerLocation::getMaxBodySize(void)
+uint32_t	ServerLocation::getMaxBodySize(void) const
 {
 	return (this->_maxBodySize);
 }
 
-IntStrMap::iterator	ServerLocation::getErrPageIttBegin(void)
+IntStrMap::const_iterator	ServerLocation::getErrPageIttBegin(void) const
 {
 	return (this->_errorPages.begin());
 }
 
-IntStrMap::iterator	ServerLocation::getErrPageIttEnd(void)
+IntStrMap::const_iterator	ServerLocation::getErrPageIttEnd(void) const
 {
 	return (this->_errorPages.end());
 }
 
-IntStrMap::iterator	ServerLocation::getRedirectionIttBegin(void)
+IntStrMap::const_iterator	ServerLocation::getRedirectionIttBegin(void) const
 {
 	return (this->_redirections.begin());
 }
 
-IntStrMap::iterator	ServerLocation::getRedirectionIttEnd(void)
+IntStrMap::const_iterator	ServerLocation::getRedirectionIttEnd(void) const
 {
 	return (this->_redirections.end());
 }
 
-http::METHOD	ServerLocation::getMethodByIndex(size_t i)
+Http::METHOD	ServerLocation::getMethodByIndex(size_t i) const
 {
 	return (this->_methodsAllowed.at(i));
 }
 
-size_t	ServerLocation::getMethodsAllowedSize(void)
+size_t	ServerLocation::getMethodsAllowedSize(void) const
 {
 	return (this->_methodsAllowed.size());
 }
 
-std::string	ServerLocation::getErrPagePath(int status)
+std::string	ServerLocation::getErrPagePath(int status) const
 {
 	//	THIS FUNCTION MUST HANDLE EDGE CASES!!!!!
-	return (this->_errorPages[status]);
+	return (this->_errorPages.at(status));
 }
 
-std::string	ServerLocation::getRedirection(int status)
+std::string	ServerLocation::getRedirection(int status) const
 {
 	//	THIS FUNCTION MUST HANDLE EDGE CASES!!!!!
-	return (this->_redirections[status]);
+	return (this->_redirections.at(status));
 }
 
-bool	ServerLocation::methodIsAllowed(Http::METHOD method)
+bool	ServerLocation::methodIsAllowed(Http::METHOD method) const
 {
 	size_t	vectorSize;
 
@@ -124,7 +129,7 @@ bool	ServerLocation::methodIsAllowed(Http::METHOD method)
 }
 
 
-//	DERIVATED CLASSES
+//	DERIVED CLASSES
 
 //		STATIC SITE
 LocationStatic::LocationStatic(void)	//	INIT BASE CLASS TOO
@@ -139,7 +144,7 @@ LocationStatic::LocationStatic(std::vector<std::string> strLocationBlock)	//	ini
 }
 */
 
-LocationStatic::LocationStatic(const LocationStatic &locationStatic)
+LocationStatic::LocationStatic(const LocationStatic &locationStatic): ServerLocation()
 {
 	if (this != &locationStatic)
 		*this = locationStatic;
@@ -151,19 +156,20 @@ LocationStatic::~LocationStatic(void)
 	return ;
 }
 
-LocationStatic	LocationStatic::&operator=(const LocationStatic &locationStatic)
+LocationStatic	&LocationStatic::operator=(const LocationStatic &locationStatic)
 {
 	this->_autoIndex = locationStatic.getAutoIndex();
 
 	return (*this);
 }
 
-bool	LocationStatic::getAutoIndex(void)
+bool	LocationStatic::getAutoIndex(void) const
 {
 	return (this->_autoIndex);
 }
 
 
+/*
 //		REVERSE PROXY (FOR FILE UPLOADS)
 LocationRevProxy::LocationRevProxy(void)
 {
@@ -218,6 +224,7 @@ LocationCGI::~LocationCGI(void)
 {
 	return ;
 }
+*/
 
 std::ostream 	&operator<<(std::ostream &out, const LocationStatic &locationStatic)
 {
@@ -227,22 +234,23 @@ std::ostream 	&operator<<(std::ostream &out, const LocationStatic &locationStati
 	out << "\tMax body size: " << locationStatic.getMaxBodySize() << std::endl;
 
 	out << "\tError pages:" << std::endl;
-	for (IntStrMap::iterator itt = locationStatic.getErrPageIttBegin(); itt != locationStatic.getErrPageIttEnd(); itt++)
+	for (IntStrMap::const_iterator itt = locationStatic.getErrPageIttBegin(); itt != locationStatic.getErrPageIttEnd(); itt++)
 		out << "\t\t" << itt->first << " " << itt->second << std::endl;
 
 	out << "\tRedirections:" << std::endl;
-	for (IntStrMap::iterator itt = locationStatic.getRedirectionIttBegin(); itt != locationStatic.getRedirectionIttEnd(); itt++)
+	for (IntStrMap::const_iterator itt = locationStatic.getRedirectionIttBegin(); itt != locationStatic.getRedirectionIttEnd(); itt++)
 		out << "\t\t" << itt->first << " " << itt->second << std::endl;
 
-	out << "Allowed methods:" << std::endl;
+	out << "\tAllowed methods:" << std::endl;
 	for (size_t i = 0; i < locationStatic.getMethodsAllowedSize(); i++)
 		out << "\t\t" << locationStatic.getMethodByIndex(i) << std::endl;
 
-	out << "\tAutoindex: " << locationStatic,getAutoIndex() << std::endl;
+	out << "\tAutoindex: " << locationStatic.getAutoIndex() << std::endl;
 
 	return (out);
 }
 
+/*
 std::ostream 	&operator<<(std::ostream &out, const LocationCGI &locationCGI)
 {
 	out << "\tLocation: " << locationCGI.getLocation() << std::endl;
@@ -266,3 +274,4 @@ std::ostream 	&operator<<(std::ostream &out, const LocationCGI &locationCGI)
 
 	return (out);
 }
+*/
