--- test_cases/compile.sh
@@ -1,1 +1,0 @@
-c++ -Wall -Wextra -Werror -std=c++98 main.cpp ServerConfig.cpp ServerLocation.cpp utils.cpp -o test
