--- test_cases/main.cpp
@@ -1,11 +1,0 @@
-#include "ServerConfig.hpp"
-#include "ServerLocation.hpp"
-
-int	main(void)
-{
-	ServerConfig	configs;
-
-	std::cout << configs << std::endl;
-
-	return (0);
-}
